Add debug data to ProcessPriceWorker

+++ Helper/ProcessPriceWorker.php
@@ -2,7 +2,7 @@
 
 class ProcessPriceWorker
 {
-    public function __construct($id, $file_name,$process_id)
+    public function __construct($id, $file_name, $process_id)
     {
         $this->id = $id;
         $this->file_name = $file_name;
@@ -11,7 +11,7 @@ class ProcessPriceWorker
 
     public function run()
     {
-        global $PHP;
+        global $PHP,$CONFIG;
 
         $handle = @fopen($this->file_name, "r");
         if (!$handle) {
@@ -29,16 +29,16 @@ class ProcessPriceWorker
         }
         fclose($handle);
 
-        $new_file = getcwd() . "/Files/" . $guid = UUID::v4() . "";
+        $new_file = getcwd() . "/Files/" . $guid = $this->id . "";
         if (!move_uploaded_file($this->file_name, $new_file)) {
             return "Невозможно переместить файл";
         }
         //exec("iconv -f windows-1251 -t utf-8 {$new_file} -o {$new_file}.csv");
         //unlink($new_file);
 
-        $outputfile = "/var/www/Medic/out.txt";
+        $outputfile = $CONFIG["PRICE_WORKER_LOG"];//"/var/www/Medic/ProcessPriceWorker.log";
         $pidfile = "pid.txt";
-        exec(sprintf("%s > %s 2>&1 & echo $! >> %s", $PHP . " " . __FILE__ . " {$this->id} {$new_file} {$this->process_id}", $outputfile, $pidfile));//.csv
+        exec(sprintf("%s > %s 2>&1 & echo $! >> %s", $PHP . " " . __FILE__ . " {$this->id} {$new_file} {$this->process_id}", $outputfile, $pidfile)); //.csv
         return "ok";
     }
 }
@@ -51,51 +51,63 @@ if (!isset($PHP)) {
     require_once 'Database.php';
     require_once 'FixSearch.php';
 
-    $db = new Database($CONFIG["DB_HOST"],$CONFIG["DB_NAME"], $CONFIG["DB_USER"], $CONFIG["DB_PASS"]);
-    $db -> Connect();
-
-    $dateExt   = new DateTime;
+    $dateExt = new DateTime;
     $cur_date = $dateExt->format("Y-m-d H:i:s");
-    $prov_data="";
+    $prov_data = "";
     $to_search = new ArrayObject();
+
+    echo "\n\n--------------------------------------------------------------------------------------";
+    echo "Start process: " . $cur_date . "\n";
+    echo "Input params: \n";
+    print_r($argv);
+
+    $db = new Database($CONFIG["DB_HOST"], $CONFIG["DB_NAME"], $CONFIG["DB_USER"], $CONFIG["DB_PASS"]);
+    $db->Connect();
 
     $filesize = filesize($argv[2]);
     $handle = @fopen($argv[2], "r");
     if (!$handle) {
+        $db->Exec("update `Uploads` set Status='Ошибка' where id='" . $argv[3] . "' ;");
         die("Error open file!");
     }
+
+    echo "Input file size:" . $filesize;
 
     if (($buffer = fgets($handle, 4096)) !== false) {
         if (substr_count($buffer, ';') != 6) {
             fclose($handle);
+            $db->Exec("update `Uploads` set Status='Ошибка' where id='" . $argv[3] . "' ;");
             die("Incorrect file format");
         }
     } else {
+        $db->Exec("update `Uploads` set Status='Ошибка' where id='" . $argv[3] . "' ;");
         die("Can't open file");
     }
-
-    //echo $buffer;
-
-    $db->Exec("update `Uploads` set Status='Подготовка 10%' where id='".$argv[3]."' ;");
-    $db->Exec("delete from `ProductsSearch` where ProductId in (select id from `Products` where ProviderId='".$argv[1]."');");
-    $db->Exec("update `Uploads` set Status='Подготовка 40%' where id='".$argv[3]."' ;");
-    $db->Exec("delete from `Products` where ProviderId='".$argv[1]."' ;");
-    $db->Exec("update `Uploads` set Status='Подготовка 70%' where id='".$argv[3]."' ;");
-    //$db->Exec("delete from `ProductsSearch` where ProductId not in (select id from `Products`);");
-
-    echo "Clean\n";
 
     $dbh;
     try {
 
-       $dbh = new PDO("mysql:host=".$CONFIG["DB_HOST"].";dbname=".$CONFIG["DB_NAME"], $CONFIG["DB_USER"], $CONFIG["DB_PASS"]);
+        echo "Structure of file: OK\n";
+        //echo $buffer;
+
+        $db->Exec("update `Uploads` set Status='Подготовка 10%' where id='" . $argv[3] . "' ;");
+        $db->Exec("delete from `ProductsSearch` where ProductId in (select id from `Products` where ProviderId='" . $argv[1] . "');");
+        $db->Exec("update `Uploads` set Status='Подготовка 40%' where id='" . $argv[3] . "' ;");
+        $db->Exec("delete from `Products` where ProviderId='" . $argv[1] . "' ;");
+        $db->Exec("update `Uploads` set Status='Подготовка 70%' where id='" . $argv[3] . "' ;");
+        //$db->Exec("delete from `ProductsSearch` where ProductId not in (select id from `Products`);");
+
+        echo "Clean before process\n";
+
+
+        $dbh = new PDO("mysql:host=" . $CONFIG["DB_HOST"] . ";dbname=" . $CONFIG["DB_NAME"], $CONFIG["DB_USER"], $CONFIG["DB_PASS"]);
 
         $sql = "SELECT Name,FullName,City,Address,Phone FROM Provider where id='{$argv[1]}' ";
         foreach ($dbh->query($sql) as $row) {
-            $prov_data = $row['Name'] . $row['FullName'] . $row['City'] . $row['Address'] .$row['Phone'] ;
+            $prov_data = $row['Name'] . $row['FullName'] . $row['City'] . $row['Address'] . $row['Phone'];
         }
 
-        echo "Prepead to process\n";
+        echo "Prepare to process\n";
         $dbh->beginTransaction();
 
         $sql = "INSERT INTO `Products`(`id`,`NumberProvider`,`Name`,`FullName`,`BasicCharacteristics`,`ProviderId`," .
@@ -103,20 +115,18 @@ if (!isset($PHP)) {
 
         $sth = $dbh->prepare($sql);
 
-        $readed=0;
-        $progress=0;
+        $readed = 0;
+        $progress = 0;
         while (($buffer = fgets($handle, 4096)) !== false) {
-            $readed+=strlen($buffer);
+            $readed += strlen($buffer);
 
-            if($progress != intval(($readed*100/$filesize)))
-            {
-                $progress = intval(($readed*100/$filesize));
-                echo "Progress:".$progress."\n";
+            if ($progress != intval(($readed * 100 / $filesize))) {
+                $progress = intval(($readed * 100 / $filesize));
+                echo "Processing " . $progress . "\n";
 
-                $db->Exec("update `Uploads` set Status='Обработка ".$progress."%' where id='".$argv[3]."' ;");
+                $db->Exec("update `Uploads` set Status='Обработка " . $progress . "%' where id='" . $argv[3] . "' ;");
             }
 
-            //echo $buffer."\n";
             $buffer = iconv("windows-1251", "UTF-8//IGNORE", $buffer);
             //$html_utf8 = mb_convert_encoding($html, "utf-8", "windows-1251");
             //$buffer = mb_strtolower($buffer, 'UTF-8');
@@ -125,55 +135,51 @@ if (!isset($PHP)) {
             $guid = UUID::v4();
 
 
-            $ss = $elem[0].$elem[1].$elem[2].$elem[3].$argv[1].$elem[4].$elem[5].$prov_data;
-            $to_search[count($to_search)]=array("id"=>$guid,"data"=>mb_strtolower($ss, 'UTF-8'));
+            $ss = $elem[0] . $elem[1] . $elem[2] . $elem[3] . $argv[1] . $elem[4] . $elem[5] . $prov_data;
+            $to_search[count($to_search)] = array("id" => $guid, "data" => mb_strtolower($ss, 'UTF-8'));
 
             $sth->execute(array(
-                $guid, $elem[0], $elem[1], $elem[2], $elem[3], $argv[1], $elem[5], $elem[6],$cur_date
+                $guid, $elem[0], $elem[1], $elem[2], $elem[3], $argv[1], $elem[5], $elem[6], $cur_date
             ));
         }
 
-
-        //$dbh->query("call FixSearch()");
         echo "Finished processing\n";
 
         $sql = "insert into `ProductsSearch` (ProductId,SearchString) values (?,?);";
 
-        echo "to_search:".count($to_search)."\n";
+        echo "to_search:" . count($to_search) . "\n";
 
-        $progress=0;
-        $col=0;
+        $progress = 0;
+        $col = 0;
         $sth = $dbh->prepare($sql);
-        foreach($to_search as $i)
-        {
+        foreach ($to_search as $i) {
             $col++;
-            if($progress != intval($col*100/count($to_search)))
-            {
-                $progress = intval($col*100/count($to_search));
-                echo "Progress:".$progress."\n";
-                $db->Exec("update `Uploads` set Status='Постобработка ".$progress."%' where id='".$argv[3]."' ;");
+            if ($progress != intval($col * 100 / count($to_search))) {
+                $progress = intval($col * 100 / count($to_search));
+                echo "Processing " . $progress . "\n";
+                $db->Exec("update `Uploads` set Status='Постобработка " . $progress . "%' where id='" . $argv[3] . "' ;");
             }
 
-            $sth->execute(array( $i["id"],$i["data"] ));
+            $sth->execute(array($i["id"], $i["data"]));
         }
         echo "Finished optimization\n";
 
         $dbh->commit();
-        //FixSearch::Go();
 
     } catch (PDOException $e) {
         $dbh->rollBack();
+        $db->Exec("update `Uploads` set Status='Ошибка' where id='" . $argv[3] . "' ;");
         die("Error!: " . $e->getMessage() . "<br/>");
     }
 
-
     if (!feof($handle)) {
+        $db->Exec("update `Uploads` set Status='Ошибка' where id='" . $argv[3] . "' ;");
         echo "Error: unexpected fgets() fail\n";
     }
-    $db->Exec("update `Uploads` set Status='Готово' where id='".$argv[3]."' ;");
+
+    $db->Exec("update `Uploads` set Status='Готово' where id='" . $argv[3] . "' ;");
     fclose($handle);
     sleep(1);
-    unlink($argv[2]);
 
-    print_r($argv);
+    if($CONFIG["DEBUG"]!=1) unlink($argv[2]);
 }
